Splits the snooze command off from the `misc` module and into its own `snooze` package. Mostly because its functionality has been expanded to the point where it warrants its own package.
The snooze command now gives the user a unique token with which
they can cancel the alarm, if desired, through the unsnooze
command. An alarm can only be cancelled with the correct token,
and only by the user (hostmask) who created the alarm in the
first place.

Additionally, all scheduled alarms are now persisted to a file
on disk and will consequently survive a bot restart. This was
previously not the case.

This commit additionally fixes some code code comments in other
module packages.