Fixes profile to save itself when the nick or nickserv passwords are changed through the `admin` plugin. This is necessary to keep the request parsing behaviour consistent after a bot restart.
Specifically, when parsing incoming requests, the bot checks
of the `Request.Target` field, refers to the bot's own nickname.
If so, this requests came from a user, through a PM, instead of
a channel. This field must then be updated to point to the sender's
name instead, so any reply ends up at the right place. The bot
would otherwise send any replies to itself.

The issue arrose when the bot had its nickname changed through
the !nick command over IRC and would afterwards be restarted
for whatever reason. On IRC, the bot would still have the newly
assigned nickname, but the bot Profile would contain the original
nickname. Consquently, messing things up left and right.

To fix this, the bot now persists any nick change to the profile.cfg
file.

Signed-off-by: Jim Teeuwen <jimteeuwen@gmail.com>