Tweaks the behaviour and output of the `stats` plugin commands to be more useful. The commands now allow a fuzzy name match, which will possibly yield more than 1 user. Since this increases the verbosity of the firston, laston and whois commands, the output of these is now always directed to the caller's PM window.
Some work still has to be done to deal with mibbit hostmasks.
We need to treat s ome hostmasks as the same user. E.g.:

    12abcdef@ircip1.mibbit.com
    12abcdef@ircip2.mibbit.com

These are the same user. The hexadecimal value before the @
symbol, is the user's IPv4 address. It translates into 4
octets: `0x12, 0xab, 0xcd, 0xef`, which yields the IP address:
`18.171.205.239`.

Signed-off-by: Jim Teeuwen <jimteeuwen@gmail.com>