Exit 1 on irc timeout or other read errors.
After an IRC timeout the bot shut down gracefully and exited with
status 0. This was a problem, as the systemd supervisor did not
restart the bot.  This fix makes the bot exit with status 1 after an
IRC timeout.

Other read errors, notably io.EOF, wil now also result in a status 1
exit.  But sometimes a read error happens when the bot is in the
process of a graceful shutdown. In this special case the error is
ignored and the bot continues the graceful shutdown.