Don't segfault when starting with stored reminders.
Fixes #14.  Planned reminders, stored in file alarms.dat, now
survive a restart. The bot does not segfault anymore when the file
alarms.dat stil contains any.

This is a bit of a nasty fix, as it assumes the bot only has exactly
one (1) connection. Elsewhere, e.g. when forking, the code is more
generic in that it does not make that assumption.

This commit also simplifies the logic around removing a reminder after
it has set off.